Moved more into SharedFaction
